Crate fs_extra

Modules

dir

This module includes additional methods for working with directories.

error

The error type for fs_extra operations on files and directories.

file

This module includes additional methods for working with files.

Structs

TransitProcess

A structure which includes information about the current status of copying or moving a directory.

Functions

copy_items

Copies a list of directories and files to another place recursively. This function will also copy the permission bits of the original files to destination files (not for directories).

copy_items_with_progress

Copies a list of directories and files to another place recursively, with information about progress. This function will also copy the permission bits of the original files to destination files (not for directories).

move_items

Moves a list of directories and files to another place recursively. This function will also copy the permission bits of the original files to destination files (not for directories).

move_items_with_progress

Moves a list of directories and files to another place recursively, with information about progress. This function will also copy the permission bits of the original files to destination files (not for directories).

remove_items

Removes a list of files or directories.