Crate fs_extra

Modules

dir

This module include extra methods for works with directories.

error

The error type for fs_extra operations with files and folder.

file

This module include extra methods for works with files.

Structs

TransitProcess

A structure which include information about the current status of the copy or move directory.

Functions

copy_items

Copies list directories and files to another place using recursive method. This function will also copy the permission bits of the original files to destionation files (not for directories).

copy_items_with_progress

Copies list directories and files to another place using recursive method, with recept information about process. This function will also copy the permission bits of the original files to destionation files (not for directories).

move_items

Moves list directories and files to another place using recursive method. This function will also copy the permission bits of the original files to destionation files (not for directories).

move_items_with_progress

Moves list directories and files to another place using recursive method, with recept information about process. This function will also copy the permission bits of the original files to destionation files (not for directories).

remove_items

Removes list files or directories.