Type Definition fs_extra::error::Result

pub type Result<T> = Result<T, Error>;

A specialized Result type for fs_extra operations.

This typedef is generally used to avoid writing out fs_extra::Error directly and is otherwise a direct mapping to Result.

#Examples

extern crate fs_extra;
use fs_extra::dir::create;

fn get_string() -> io::Result<()> {

    create("test_dir")?;

    Ok(())
}